se agrega servicio soap

+++ src/main/java/com/programalol/usuario/domain/model/Resultado.java
@@ -14,8 +14,8 @@ import java.util.Date;
 public class Resultado {
 
     private String id;
-    private int valor1;
-    private int valor2;
-    private int resultado;
+    private double valor1;
+    private double valor2;
+    private double resultado;
     private Date fechaOperacion;
 }
