+++ src/main/java/com/programalol/usuario/infraestructure/driver_adapters/nosql_repository/ResultadoDataRepository.java
@@ -7,6 +7,6 @@ import reactor.core.publisher.Mono;
 import java.util.Date;
 
 public interface ResultadoDataRepository extends ReactiveCrudRepository<ResultadoData, String> {
-    Mono<Void> deleteByValor1(int valor1);
+    Mono<Void> deleteByValor1(double valor1);
     Mono<Void> deleteByFechaOperacionBefore(Date fechaOperacion);
 }
